+++ cg/cli/workflow/microsalt/base.py
@@ -6,13 +6,16 @@ from pathlib import Path
 from typing import Any
 
 import click
+import datetime as dt
 
 from cg.apps.hermes.hermes_api import HermesApi
 from cg.apps.housekeeper.hk import HousekeeperAPI
 from cg.apps.lims import LimsAPI
+from cg.apps.tb import TrailblazerAPI
+from cg.apps.vogue import VogueAPI
 from cg.constants import EXIT_FAIL, EXIT_SUCCESS, Pipeline
 from cg.meta.workflow.microsalt import MicrosaltAnalysisAPI
-from cg.store import Store
+from cg.store import Store, models
 
 LOG = logging.getLogger(__name__)
 
@@ -39,7 +42,7 @@ ARGUMENT_UNIQUE_IDENTIFIER = click.argument("unique_id", required=True, type=cli
 
 @click.group(invoke_without_command=True)
 @click.pass_context
-def microsalt(context: click.Context):
+def microsalt(context: click.Context) -> None:
     """Microbial workflow"""
     if context.invoked_subcommand is None:
         click.echo(context.get_help())
@@ -48,6 +51,7 @@ def microsalt(context: click.Context):
         db=Store(context.obj["database"]),
         hk_api=HousekeeperAPI(context.obj),
         lims_api=LimsAPI(context.obj),
+        trailblazer_api=TrailblazerAPI(context.obj),
         hermes_api=HermesApi(context.obj),
         config=context.obj["microsalt"],
     )
@@ -59,8 +63,8 @@ def microsalt(context: click.Context):
 @OPTION_SAMPLE
 @ARGUMENT_UNIQUE_IDENTIFIER
 @click.pass_context
-def link(context: click.Context, ticket: bool, sample: bool, unique_id: str):
-    """Link microbial FASTQ files for a SAMPLE_ID"""
+def link(context: click.Context, ticket: bool, sample: bool, unique_id: str) -> None:
+    """Link microbial FASTQ files to dedicated analysis folder for a given case, ticket or sample"""
 
     microsalt_analysis_api: MicrosaltAnalysisAPI = context.obj["microsalt_analysis_api"]
 
@@ -82,8 +86,10 @@ def link(context: click.Context, ticket: bool, sample: bool, unique_id: str):
 @OPTION_SAMPLE
 @ARGUMENT_UNIQUE_IDENTIFIER
 @click.pass_context
-def config_case(context: click.Context, dry_run: bool, ticket: bool, sample: bool, unique_id: str):
-    """ Create a config file on case level for microSALT """
+def config_case(
+    context: click.Context, dry_run: bool, ticket: bool, sample: bool, unique_id: str
+) -> None:
+    """ Create a config file for a case or a sample analysis in microSALT """
 
     microsalt_analysis_api: MicrosaltAnalysisAPI = context.obj["microsalt_analysis_api"]
 
@@ -131,8 +137,8 @@ def run(
     ticket: bool,
     sample: bool,
     unique_id: Any,
-):
-    """ Start microSALT with an order_id """
+) -> None:
+    """ Start microSALT workflow by providing case, ticket or sample id """
 
     microsalt_analysis_api: MicrosaltAnalysisAPI = context.obj["microsalt_analysis_api"]
 
@@ -162,14 +168,25 @@ def run(
         ]
     microsalt_analysis_api.process.run_command(parameters=analyse_command, dry_run=dry_run)
 
-    if not sample_id and not dry_run:
-        microsalt_analysis_api.set_statusdb_action(case_id=case_id, action="running")
+    if sample_id or dry_run:
+        return
+
+    microsalt_analysis_api.set_statusdb_action(case_id=case_id, action="running")
+    try:
+        microsalt_analysis_api.submit_trailblazer_analysis(case_id=case_id)
+    except Exception as e:
+        LOG.warning(
+            "Trailblazer warning: Could not track analysis progress for case %s! %s",
+            case_id,
+            e.__class__.__name__,
+        )
 
 
 @microsalt.command()
 @ARGUMENT_UNIQUE_IDENTIFIER
 @click.pass_context
-def store(context: click.Context, unique_id: str):
+def store(context: click.Context, unique_id: str) -> None:
+    """Store microSALT results in Housekeeper for given case"""
     microsalt_analysis_api: MicrosaltAnalysisAPI = context.obj["microsalt_analysis_api"]
 
     case_obj = microsalt_analysis_api.db.family(unique_id)
@@ -195,7 +212,10 @@ def store(context: click.Context, unique_id: str):
 @OPTION_TICKET
 @OPTION_SAMPLE
 @click.pass_context
-def start(context: click.Context, ticket: bool, sample: bool, unique_id: str, dry_run: bool):
+def start(
+    context: click.Context, ticket: bool, sample: bool, unique_id: str, dry_run: bool
+) -> None:
+    """Start whole microSALT workflow by providing case, ticket or sample id"""
     LOG.info("Starting Microsalt workflow for %s", unique_id)
     context.invoke(link, ticket=ticket, sample=sample, unique_id=unique_id)
     context.invoke(config_case, ticket=ticket, sample=sample, unique_id=unique_id, dry_run=dry_run)
@@ -205,7 +225,8 @@ def start(context: click.Context, ticket: bool, sample: bool, unique_id: str, dr
 @microsalt.command("store-available")
 @OPTION_DRY_RUN
 @click.pass_context
-def store_available(context: click.Context, dry_run: bool):
+def store_available(context: click.Context, dry_run: bool) -> None:
+    """Store all finished analyses in Housekeeper"""
     microsalt_analysis_api: MicrosaltAnalysisAPI = context.obj["microsalt_analysis_api"]
     exit_code: int = EXIT_SUCCESS
 
@@ -225,7 +246,8 @@ def store_available(context: click.Context, dry_run: bool):
 @microsalt.command("start-available")
 @OPTION_DRY_RUN
 @click.pass_context
-def start_available(context: click.Context, dry_run: bool):
+def start_available(context: click.Context, dry_run: bool) -> None:
+    """Start whole microSALT workflow for all newly sequenced cases"""
     microsalt_analysis_api: MicrosaltAnalysisAPI = context.obj["microsalt_analysis_api"]
     exit_code: int = EXIT_SUCCESS
     for case_obj in microsalt_analysis_api.db.cases_to_analyze(pipeline=Pipeline.MICROSALT):
@@ -243,4 +265,83 @@ def start_available(context: click.Context, dry_run: bool):
             exit_code = EXIT_FAIL
 
     if exit_code:
+        raise click.Abort
+
+
+@microsalt.command("upload-analysis-vogue")
+@OPTION_DRY_RUN
+@ARGUMENT_UNIQUE_IDENTIFIER
+@click.pass_context
+def upload_analysis_vogue(context: click.Context, unique_id: str, dry_run: bool) -> None:
+    """Upload the trending report for latest analysis of given case_id to Vogue"""
+
+    vogue_api = VogueAPI(context.obj)
+    microsalt_analysis_api: MicrosaltAnalysisAPI = context.obj["microsalt_analysis_api"]
+    case_obj = microsalt_analysis_api.db.family(unique_id)
+    if not case_obj or not case_obj.analyses:
+        LOG.error("No analysis available for %s", unique_id)
+        raise click.Abort
+
+    samples_string = ",".join(str(link_obj.sample.internal_id) for link_obj in case_obj.links)
+    microsalt_version = microsalt_analysis_api.get_microsalt_version()
+
+    if dry_run:
+        LOG.info(
+            "Would have loaded case %s, with samples %s, analyzed with pipeline version %s",
+            unique_id,
+            samples_string,
+            microsalt_version,
+        )
+        return
+
+    analysis_result_file = microsalt_analysis_api.hk.get_files(
+        bundle=unique_id, tags=["vogue"]
+    ).first()
+    if not analysis_result_file:
+        LOG.error("Vogue upload file not found in Housekeeper for case %s", unique_id)
+        raise click.Abort
+
+    vogue_load_args = {
+        "samples": samples_string,
+        "analysis_result_file": analysis_result_file.full_path,
+        "analysis_case_name": unique_id,
+        "analysis_type": "microsalt",
+        "analysis_workflow_name": "microsalt",
+        "analysis_workflow_version": microsalt_version,
+        "case_analysis_type": "microsalt",
+    }
+    vogue_api.load_bioinfo_raw(load_bioinfo_inputs=vogue_load_args)
+    vogue_api.load_bioinfo_process(load_bioinfo_inputs=vogue_load_args, cleanup_flag=False)
+    vogue_api.load_bioinfo_sample(load_bioinfo_inputs=vogue_load_args)
+    case_obj.analyses[0].uploaded_at = dt.datetime.now()
+    microsalt_analysis_api.db.commit()
+    LOG.info("Successfully uploaded latest analysis data for case %s to Vogue!", unique_id)
+
+
+@microsalt.command("upload-latest-analyses-vogue")
+@OPTION_DRY_RUN
+@click.pass_context
+def upload_vogue_latest(context: click.Context, dry_run: bool) -> None:
+    """Upload the trending reports for all un-uploaded latest analyses to Vogue"""
+
+    EXIT_CODE: int = EXIT_SUCCESS
+    microsalt_analysis_api: MicrosaltAnalysisAPI = context.obj["microsalt_analysis_api"]
+    latest_analyses = list(
+        microsalt_analysis_api.db.latest_analyses()
+        .filter(models.Analysis.pipeline == Pipeline.MICROSALT)
+        .filter(models.Analysis.uploaded_at.is_(None))
+    )
+    for analysis in latest_analyses:
+        unique_id = analysis.family.internal_id
+        try:
+            context.invoke(upload_analysis_vogue, unique_id=unique_id, dry_run=dry_run)
+        except Exception as e:
+            LOG.error(
+                "Could not upload data for %s to vogue, exception %s",
+                unique_id,
+                e.__class__.__name__,
+            )
+            EXIT_CODE = EXIT_FAIL
+
+    if EXIT_CODE:
         raise click.Abort