+++ tests/cli/upload/conftest.py
@@ -8,11 +8,13 @@ import pytest
 
 from cg.apps.gt import GenotypeAPI
 from cg.apps.housekeeper.hk import HousekeeperAPI
-from cg.apps.scout.scout_load_config import ScoutLoadConfig
 from cg.apps.scout.scoutapi import ScoutAPI
-from cg.meta.upload.scoutapi import UploadScoutAPI
+from cg.meta.upload.scout.scout_load_config import ScoutLoadConfig
+from cg.meta.upload.scout.scoutapi import UploadScoutAPI
 from cg.meta.workflow.mip import MipAnalysisAPI
 from cg.store import Store, models
+from tests.meta.upload.scout.conftest import fixture_mip_load_config
+from tests.mocks.hk_mock import MockHousekeeperAPI
 from tests.mocks.madeline import MockMadelineAPI
 
 LOG = logging.getLogger(__name__)
@@ -77,7 +79,7 @@ def fixture_upload_genotypes_hk_api(
     return real_housekeeper_api
 
 
-@pytest.yield_fixture(name="upload_genotypes_context")
+@pytest.fixture(name="upload_genotypes_context")
 def fixture_upload_genotypes_context(
     upload_genotypes_hk_api: HousekeeperAPI, genotype_api: GenotypeAPI, analysis_store_trio: Store
 ) -> dict:
@@ -129,10 +131,11 @@ def fixture_vogue_cli_context(vogue_api) -> dict:
 
 
 @pytest.fixture(scope="function", name="upload_scout_api")
-def fixture_upload_scout_api(housekeeper_api):
+def fixture_upload_scout_api(housekeeper_api: MockHousekeeperAPI, mip_load_config: ScoutLoadConfig):
     """Return a upload scout api"""
     api = MockScoutUploadApi()
     api.housekeeper = housekeeper_api
+    api.config = mip_load_config
 
     return api
 
@@ -185,9 +188,10 @@ class MockScoutUploadApi(UploadScoutAPI):
         self.housekeeper = None
         self.madeline_api = MockMadelineAPI()
         self.analysis = MockAnalysisApi()
-        self.config = {}
+        self.config = ScoutLoadConfig()
         self.file_exists = False
         self.lims = MockLims()
+        self.missing_mandatory_field = False
 
     @pytest.fixture(autouse=True)
     def _request_analysis(self, analysis_store_single_case):
@@ -195,12 +199,12 @@ class MockScoutUploadApi(UploadScoutAPI):
 
     def generate_config(self, analysis_obj, **kwargs):
         """Mock the generate config"""
-        if self.mock_generate_config:
-            return self.config
+        if self.missing_mandatory_field:
+            self.config.vcf_snv = None
 
-        return super().generate_config(analysis_obj, **kwargs)
+        return self.config
 
-    def save_config_file(self, scout_config, file_path):
+    def save_config_file(self, upload_config, file_path):
         """docstring for save_config_file"""
         return